# C# performance: pool and reuse vertex objects to reduce allocations when using Clear()